+++ api/controllers/api/admin/ProductController.js
@@ -6,8 +6,8 @@ module.exports = {
       const modelName = req.options.controller.split("/").reverse()[0];
       let result;
       if (serverSidePaging) {
-        const include = [ ProductDescription, Supplier, { model: Category, include: CategoryDescription  }];
-        result = await PagingService.process({query, modelName, include});
+        // const include = [ ProductDescription, Supplier, { model: Category, include: CategoryDescription  }];
+        result = await PagingService.process({query, modelName});
       } else {
         const items = await sails.models[modelName].findAll({
           include: [ ProductDescription, Supplier, { model: Category, include: CategoryDescription  }]