+++ api/models/Product.js
@@ -49,6 +49,12 @@ module.exports = {
       defaultValue: 'tw',
     },
 
+    origin: {
+      type: Sequelize.STRING(128),
+      allowNull: false,
+      defaultValue: 'tw',
+    },
+
     quantity: {
       type: Sequelize.INTEGER(4),
       allowNull: false,
@@ -72,10 +78,21 @@ module.exports = {
       defaultValue: '0.0000',
     },
 
+    taxRate: {
+      type: Sequelize.FLOAT,
+      allowNull: false,
+      defaultValue: sails.config.taxrate,
+    },
+
     points: {
       type: Sequelize.INTEGER(8),
       allowNull: false,
       defaultValue: 0,
+    },
+
+    precautions: {
+      type: Sequelize.STRING(255),
+      allowNull: true,
     },
 
     // taxClassId: {
@@ -211,6 +228,31 @@ module.exports = {
       }
     },
 
+    suppliersId: {
+      type: Sequelize.VIRTUAL,
+      get: function(){
+        try{
+          const tihsSuppliers = this.getDataValue('Suppliers');
+          const suppliers = tihsSuppliers ? tihsSuppliers.map((supplier) => supplier.id) : [];
+          return suppliers;
+        } catch(e){
+          sails.log.error(e);
+        }
+      }
+    },
+    suppliersNameArray: {
+      type: Sequelize.VIRTUAL,
+      get: function(){
+        try{
+          const tihsSuppliers = this.getDataValue('Suppliers');
+          const suppliers = tihsSuppliers ? tihsSuppliers.map((supplier) => supplier.name) : [];
+          return suppliers;
+        } catch(e){
+          sails.log.error(e);
+        }
+      }
+    },
+
     // stockStatusId: {
     //   type: Sequelize.INTEGER(11),
     //   allowNull: false,
@@ -251,7 +293,10 @@ module.exports = {
     Product.hasMany(ProductOptionValue);
     Product.hasMany(ProductImage);
     Product.belongsTo(Image);
-    Product.belongsTo(Supplier);
+    // Product.belongsTo(Supplier);
+    Product.belongsToMany(Supplier, {
+      through: 'ProductSupplier'
+    });
 
     Product.belongsToMany(Category, {
       through: 'ProductCategory',