+++ react-app-ship/src/views/ShipCompletedListView/components/CardBodyExpend.js
@@ -37,7 +37,7 @@ function CardBodyExpend(props) {
         <div className='main-title'>
           <FlatButton
             className='btn-print'
-            label='列印出貨單'
+            label='列印出貨明細'
             onClick={props.handleBtnPrint}
           />
         </div>