+++ migration/b2b/changelog.md
@@ -22,6 +22,12 @@
 ### SupplierShipOrder 欄位新增
 * 新增 shippingEmail, shippingTelephone, tax , totalIncludeTax
 
-v3
+
+## v3
+
 ### 更改 MenuItem 項目
 隱藏 orderpayment, orderpaymenthistory, orderpaymentstatus
+
+### Order, OrderProducts, SupplierShipOrder, SupplierShipOrderProduct 增加欄位
+* 增加 deletedAt，搭配 model paranoid 選項
+